created separate module for steals in simulator
