+++ triple_triple/simulator/simulate_plays.py
@@ -13,6 +13,11 @@ from triple_triple.simulator.blocks_affect_for_sim import (
     blocks_affect_on_taking_shot,
     blocks_affect_on_making_shot
 )
+from triple_triple.simulator.steals_affect_for_sim import (
+    steals_affect_on_turnover,
+    update_turnover_params,
+    check_turnover_is_steal
+)
 from triple_triple.simulator.rebounds_for_sim import update_rebound_params
 
 
@@ -264,30 +269,6 @@ def update_pass_params(game_class, off_game_idx, has_ball_class):
     has_ball_class.passes += 1
 
 
-# def dist_two_points(point1, point2):
-#     return np.sqrt((point1[0] - point2[0]) ** 2 + (point1[1] - point2[1]) ** 2)
-# 
-# 
-# def get_closest_to_ball(players_dict, ball_class):
-#     player_id_list = []
-#     dist_to_ball = []
-# 
-#     for player_class in players_dict.values():
-#         player_id_list.append(player_class.player_id)
-#         dist_to_ball.append(
-#             dist_two_points(
-#                 point1=ball_class.court_coord,
-#                 point2=player_class.court_coord
-#             )
-#         )
-#     dist_to_ball = np.array(dist_to_ball)
-#     min_dist_idx = np.where(
-#         dist_to_ball == dist_to_ball.min()
-#     )[0]
-# 
-#     return [player_id_list[i] for i in min_dist_idx]
-
-
 def player_blocked_possession_switch(defender_class, teams_list, shooting_side_list):
     # switch team possession
     teams_list, shooting_side_list = switch_possession_params(
@@ -302,35 +283,6 @@ def player_blocked_possession_switch(defender_class, teams_list, shooting_side_l
         has_ball_class=defender_class
     )
     return defender_class.player_id, teams_list, shooting_side_list
-
-
-# def who_gets_rebound(
-#     teams_list,
-#     ball_class
-# ):
-#     players_offense_dict, players_defense_dict = teams_list[0], teams_list[1]
-#     closest_off_players = get_closest_to_ball(
-#         players_dict=players_offense_dict,
-#         ball_class=ball_class
-#     )
-#     closest_def_players = get_closest_to_ball(
-#         players_dict=players_defense_dict,
-#         ball_class=ball_class
-#     )
-#     off_reb_prob = [players_offense_dict[player].off_rebounds_poss for player in closest_off_players]
-#     def_reb_prob = [players_defense_dict[player].def_rebounds_poss for player in closest_def_players]
-# 
-#     # scale combined probability array
-#     prob_array = np.array(off_reb_prob + def_reb_prob) / sum(off_reb_prob + def_reb_prob)
-# 
-#     rebounder_id = np.random.choice(
-#         a=closest_off_players + closest_def_players,
-#         p=prob_array
-#     )
-#     try:
-#         return players_defense_dict[rebounder_id]
-#     except:
-#         return players_offense_dict[rebounder_id]
 
 
 def update_rebound_missed_shot(teams_list, game_class, shooting_side_list):
@@ -353,53 +305,6 @@ def update_rebound_missed_shot(teams_list, game_class, shooting_side_list):
     )
 
     return rebounder_class.player_id, teams_list, shooting_side_list
-
-
-# def update_rebound_params(teams_list, game_class):
-#     rebounder_class = who_gets_rebound(
-#         teams_list=teams_list,
-#         ball_class=ball_class,
-#     )
-# 
-#     rebounder_class.has_possession = True
-# 
-#     if rebounder_class.on_defense:
-#         rebounder_class.def_rebounds += 1
-#         game_class.def_rebounds[rebounder_class.game_idx] += 1
-# 
-#     else:
-#         rebounder_class.off_rebounds += 1
-#         game_class.off_rebounds[rebounder_class.game_idx] += 1
-# 
-#     return rebounder_class
-
-
-def steals_affect_on_turnover(turnover_prob, steals_prob):
-    return turnover_prob + steals_prob
-
-
-def update_turnover_params(has_ball_class, game_class, off_game_idx):
-    game_class.turnovers[off_game_idx] += 1
-    has_ball_class.turnovers += 1
-    has_ball_class.has_possession = False
-
-
-def check_turnover_is_steal(
-    defender_class,
-    game_class,
-):
-    # determine if turnover action is steal
-    turnover_steal = np.random.choice(
-        a=[True, False],
-        p=[defender_class.steals_poss, 1 - defender_class.steals_poss]
-    )
-
-    if turnover_steal:
-        defender_class.steals += 1
-        game_class.steals[defender_class.game_idx] += 1
-        defender_class.has_possession = True
-
-    return turnover_steal
 
 
 def update_ball_stolen(defender_class, shooting_side):
